Smart Expense Manager/goals/serializers.py
from rest_framework import serializers
from .models import Goal

class GoalSerializer(serializers.ModelSerializer):
    progress_percentage = serializers.ReadOnlyField()

    class Meta:
        model = Goal
        fields = '__all__'
        read_only_fields = ('user', 'created_at', 'progress_percentage')
