Smart Expense Manager/requirements.txt
Django>=5.0
Pillow>=10.0
pandas>=2.0
scikit-learn>=1.3
numpy>=1.24
python-dotenv>=1.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.0
